requirements.txt
asyncpg==0.30.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
disnake==2.10.1
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
idna==3.10
multidict==6.1.0
nats-py==2.9.0
propcache==0.2.1
protobuf==5.29.3
setuptools==75.8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
yarl==1.18.3